+++ test/acceptance/dataviews/histogram.js
@@ -888,7 +888,6 @@ describe('histogram-dates: aggregation input value', function() {
     });
 });
 
-
 describe('histogram-dates: timestamp starts at epoch', function() {
 
     afterEach(function(done) {
@@ -953,6 +952,224 @@ describe('histogram-dates: timestamp starts at epoch', function() {
 
             assert.equal(timestamp_start, 0);
             assert.equal(aggregation, 'month');
+
+            done();
+        });
+    });
+});
+
+describe('histogram-dates: trunc timestamp for each bin respecting user\'s timezone', function() {
+
+    afterEach(function(done) {
+        if (this.testClient) {
+            this.testClient.drain(done);
+        } else {
+            done();
+        }
+    });
+
+    var mapConfig = createMapConfig(
+        [
+            {
+                type: "cartodb",
+                options: {
+                    source: {
+                        id: "a0"
+                    },
+                    cartocss: "#points { marker-width: 10; marker-fill: red; }",
+                    cartocss_version: "2.3.0"
+                }
+            }
+        ],
+        {
+            timezone_epoch_histogram: {
+                source: {
+                    id: 'a0'
+                },
+                type: 'histogram',
+                options: {
+                    column: 'd',
+                    aggregation: 'auto'
+                }
+            },
+            timezone_epoch_histogram_tz: {
+                source: {
+                    id: 'a1'
+                },
+                type: 'histogram',
+                options: {
+                    column: 'd',
+                    aggregation: 'auto'
+                }
+            }
+        },
+        [
+            {
+                id: 'a0',
+                type: 'source',
+                params: {
+                    query: [
+                        'select null::geometry the_geom_webmercator, date AS d',
+                        'from generate_series(',
+                            '\'1970-01-01 00:00:00\'::timestamp,',
+                            '\'1970-01-01 01:59:00\'::timestamp,',
+                            ' \'1 minute\'::interval',
+                        ') date'
+                    ].join(' ')
+                }
+            },
+            {
+                id: 'a1',
+                type: 'source',
+                params: {
+                    query: [
+                        'select null::geometry the_geom_webmercator, date AS d',
+                        'from generate_series(',
+                            '\'1970-01-01 00:00:00\'::timestamptz,',
+                            '\'1970-01-01 01:59:00\'::timestamptz,',
+                            ' \'1 minute\'::interval',
+                        ') date'
+                    ].join(' ')
+                }
+            }
+        ]
+    );
+
+    var dateHistogramsUseCases = [{
+        desc: 'supporting timestamp with offset',
+        dataviewId: 'timezone_epoch_histogram_tz'
+    }, {
+        desc: 'supporting timestamp without offset',
+        dataviewId: 'timezone_epoch_histogram'
+    }];
+
+    dateHistogramsUseCases.forEach(function (test) {
+        it('should return histogram with two buckets ' + test.desc , function(done) {
+            this.testClient = new TestClient(mapConfig, 1234);
+
+            const override = {
+                aggregation: 'day',
+                offset: '-3600'
+            };
+
+            this.testClient.getDataview(test.dataviewId, override, function(err, dataview) {
+                assert.ifError(err);
+
+                var OFFSET_IN_MINUTES = -1 * 60; // GMT-01
+                var initialTimestamp = '1969-12-31T00:00:00-01:00';
+                var binsStartInMilliseconds = dataview.bins_start * 1000;
+                var binsStartFormatted = moment.utc(binsStartInMilliseconds)
+                    .utcOffset(OFFSET_IN_MINUTES)
+                    .format();
+                assert.equal(binsStartFormatted, initialTimestamp);
+
+                dataview.bins.forEach(function (bin, index) {
+                    var binTimestampExpected = moment.utc(initialTimestamp)
+                        .utcOffset(OFFSET_IN_MINUTES)
+                        .add(index, override.aggregation)
+                        .format();
+                    var binsTimestampInMilliseconds = bin.timestamp * 1000;
+                    var binTimestampFormatted = moment.utc(binsTimestampInMilliseconds)
+                        .utcOffset(OFFSET_IN_MINUTES)
+                        .format();
+
+                    assert.equal(binTimestampFormatted, binTimestampExpected);
+                    assert.ok(bin.timestamp <= bin.min, 'bin timestamp < bin min: ' + JSON.stringify(bin));
+                    assert.ok(bin.min <= bin.max, 'bin min < bin max: ' + JSON.stringify(bin));
+                });
+
+                done();
+            });
+        });
+    });
+});
+
+
+describe('histogram: be able to override with aggregation for histograms instantiated w/o aggregation', function() {
+
+    afterEach(function(done) {
+        if (this.testClient) {
+            this.testClient.drain(done);
+        } else {
+            done();
+        }
+    });
+
+    var mapConfig = createMapConfig(
+        [
+            {
+                type: "cartodb",
+                options: {
+                    source: {
+                        id: "a0"
+                    },
+                    cartocss: "#points { marker-width: 10; marker-fill: red; }",
+                    cartocss_version: "2.3.0"
+                }
+            }
+        ],
+        {
+            timezone_epoch_histogram: {
+                source: {
+                    id: 'a0'
+                },
+                type: 'histogram',
+                options: {
+                    column: 'd',
+                }
+            }
+        },
+        [
+            {
+                id: 'a0',
+                type: 'source',
+                params: {
+                    query: [
+                        'select null::geometry the_geom_webmercator, date AS d',
+                        'from generate_series(',
+                            '\'1970-01-01 00:00:00\'::timestamp,',
+                            '\'1970-01-01 01:59:00\'::timestamp,',
+                            ' \'1 minute\'::interval',
+                        ') date'
+                    ].join(' ')
+                }
+            }
+        ]
+    );
+
+    it('should apply aggregation to the histogram', function(done) {
+        this.testClient = new TestClient(mapConfig, 1234);
+
+        const override = {
+            aggregation: 'day',
+            offset: '-3600'
+        };
+
+        this.testClient.getDataview('timezone_epoch_histogram', override, function(err, dataview) {
+            assert.ifError(err);
+
+            var OFFSET_IN_MINUTES = -1 * 60; // GMT-01
+            var initialTimestamp = '1969-12-31T00:00:00-01:00';
+            var binsStartInMilliseconds = dataview.bins_start * 1000;
+            var binsStartFormatted = moment.utc(binsStartInMilliseconds)
+                .utcOffset(OFFSET_IN_MINUTES)
+                .format();
+            assert.equal(binsStartFormatted, initialTimestamp);
+
+            dataview.bins.forEach(function (bin, index) {
+                var binTimestampExpected = moment.utc(initialTimestamp)
+                    .utcOffset(OFFSET_IN_MINUTES)
+                    .add(index, override.aggregation)
+                    .format();
+                var binsTimestampInMilliseconds = bin.timestamp * 1000;
+                var binTimestampFormatted = moment.utc(binsTimestampInMilliseconds)
+                    .utcOffset(OFFSET_IN_MINUTES)
+                    .format();
+
+                assert.equal(binTimestampFormatted, binTimestampExpected);
+                assert.ok(bin.timestamp <= bin.min, 'bin timestamp < bin min: ' + JSON.stringify(bin));
+                assert.ok(bin.min <= bin.max, 'bin min < bin max: ' + JSON.stringify(bin));
+            });
 
             done();
         });