+++ NEWS.md
@@ -3,6 +3,14 @@
 ## 3.10.2
 Released 2017-mm-dd
 
+Announcements:
+ - Allow to override with any aggregation for histograms instantiated w/o aggregation.
+
+Bug fixes:
+ - Apply timezone after truncating the minimun date for each bin to calculate timestamps in time-series.
+ - Support timestamp with timezones to calculate the number of bins in time-series.
+ - Fixed issue related to name collision while building time-series query.
+
 
 ## 3.10.1
 Released 2017-08-04